# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam Project",
    "description": "Documentation Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/roles/": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "Returns all roles. Only admin can view",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "Returns one role which has an id equals to the provided id in the url path. Only admin can view",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/roles/create": {
      "post": {
        "tags": [
          "Roles"
        ],
        "description": "Create a new role. Only admin has access.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains a roleName",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/roles/update/{id}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "description": "Updates a role. id of the target role must be provided in the url path. Only admin has access",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains a field newName",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/roles/delete/{id}": {
      "delete": {
        "tags": [
          "Roles"
        ],
        "description": "Deletes a specific role. Role's id must be provided in the url path. Only admin!",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Returns all the products",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Returns only one product for id provided in url path",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/create": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Adds a new product. Only admin user has access",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains credentials such as name, imgUrl, description, price, countInStock, brandName and categoryName.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/update/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Change/updates a specific product for provided id in the url path. Only admin!",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which must contain following credentials:   name, imgUrl, description, price, countInStock, brandName, categoryName.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/restore/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Restore a specific product for provided id in the url path. Only admin!",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/delete/{id}": {
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Deletes a specific product for the provided id in url path. Only admin!",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateRole": {
      "type": "object",
      "properties": {
        "roleName": {
          "type": "string",
          "example": "User"
        }
      },
      "required": [
        "roleName"
      ]
    },
    "UpdateRole": {
      "type": "object",
      "properties": {
        "newName": {
          "type": "string",
          "example": "Admin"
        }
      },
      "required": [
        "newName"
      ]
    },
...
    "CreateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "test-product"
        },
        "imgUrl": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-apple-tv.png"
        },
        "description": {
          "type": "string",
          "example": "a test product"
        },
        "price": {
          "type": "number",
          "example": 1200
        },
        "countInStock": {
          "type": "number",
          "example": 7
        },
        "brandName": {
          "type": "string",
          "example": "Apple"
        },
        "categoryName": {
          "type": "string",
          "example": "phones"
        }
      },
      "required": [
        "name",
        "imgUrl",
        "description",
        "price",
        "countInStock",
        "brandName",
        "categoryName"
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "newName": {
          "type": "string",
          "example": "new test-product"
        },
        "newImgUrl": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-apple-tv.png"
        },
        "newDescription": {
          "type": "string",
          "example": "a new test product"
        },
        "newPrice": {
          "type": "number",
          "example": 2200
        },
        "newCountInStock": {
          "type": "number",
          "example": 8
        },
        "newBrandName": {
          "type": "string",
          "example": "Samsung"
        },
        "newCategoryName": {
          "type": "string",
          "example": "tablets"
        }
      },
      "required": [
        "newName",
        "newImgUrl",
        "newDescription",
        "newPrice",
        "newCountInStock",
        "newBrandName",
        "newCategoryName"
      ]
    },
...
  }
}